Updated to BigInteger

+++ HailstoneNumber.cs
@@ -1,11 +1,12 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 
 namespace ThreeXplusOne
 {
     public class HailstoneNumber
     {
-        public HailstoneNumber(ulong number)
+        public HailstoneNumber(BigInteger number)
         {
             Number = number;
             NumberSet.Add(this);
@@ -17,12 +18,12 @@ namespace ThreeXplusOne
             }
         }
 
-        public ulong MaxHeight => NumberSet.Max(n => n.Number);
+        public BigInteger MaxHeight => NumberSet.Max(n => n.Number);
         public int StoppingTime => NumberSet.Count;
-        public ulong Number { get; }
+        public BigInteger Number { get; }
         public List<HailstoneNumber> NumberSet { get; } = new();
 
-        private ulong NextNumber() => Number % 2 == 0 ? Number / 2 : Number * 3 + 1;
+        private BigInteger NextNumber() => Number % 2 == 0 ? Number / 2 : Number * 3 + 1;
         private bool FloorHit() => Number == 1;
         public string NumberSetString() => string.Join(@", ", NumberSet.Select(ns => ns.Number));
